+++ datatig/repository_access.py
@@ -14,10 +14,24 @@ class RepositoryAccess(ABC):
     def get_contents_of_file(self, file_name: str) -> str:
         return ""
 
+    @abstractmethod
+    def has_file(self, file_name: str) -> bool:
+        return False
+
 
 class RepositoryAccessLocalFiles(RepositoryAccess):
     def __init__(self, source_dir: str):
         self._source_dir = source_dir
+
+    def has_file(self, file_name: str) -> bool:
+        full_start_dir = os.path.abspath(self._source_dir)
+        for path, subdirs, files in os.walk(self._source_dir):
+            for name in files:
+                full_filename = os.path.abspath(os.path.join(path, name))
+                found_filename = full_filename[len(full_start_dir) + 1 :]
+                if found_filename == file_name:
+                    return True
+        return False
 
     def list_files_in_directory(self, directory_name: str) -> list:
         out = []
@@ -48,6 +62,20 @@ class RepositoryAccessLocalGit(RepositoryAccess):
 
     def set_ref(self, ref: str) -> None:
         self._ref = ref if ref else "HEAD"
+
+    def has_file(self, file_name: str) -> bool:
+        process = subprocess.Popen(
+            ["git", "ls-tree", "-r", self._ref],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            cwd=self._source_dir,
+        )
+        stdout, stderr = process.communicate()
+        for line in stdout.decode("utf-8").strip().split("\n"):
+            path_relative_to_repo = line.split("\t")[-1]
+            if path_relative_to_repo == file_name:
+                return True
+        return False
 
     def list_files_in_directory(self, directory_name: str) -> list:
         out = []